+++ Changelog.md
@@ -5,6 +5,7 @@
 - Multiple AviSynth.dlls were included due to bug in release script. (stax76, 44vince44 [#407](/../../issues/407))
 - Apple QuickTime library removed.
 - The apps dialog has a feature to copy the path to the clipboard. (stax76)
+- The apps dialog has a feature to edit the user PATH environment variable. (stax76 [#305](/../../issues/305))
 - mpv.net 5.4.8.6